arreglo del siderbar/header responsitive
