+++ src/app/app.component.html
@@ -8,23 +8,3 @@
     <router-outlet></router-outlet>
   </div>
  
-
-
-<!-- <router-outlet></router-outlet> -->
-
-<!-- <div class="wrapper">
-    <header class="right-section">
-        <app-header></app-header>
-    </header>
-    <div class="sidenav" [ngClass]="isSideNavCollapsed ? 'sidenav-collapsed' : ''">
-        <app-sidebar (onToggleSideNav)="onToggleSideNav($event)"></app-sidebar>
-    </div>
-    <div [ngClass]="isSideNavCollapsed ? 'content-collapsed' : 'content-expanded'">
-        <div>
-            <app-map *ngIf="showMap"></app-map>
-        </div>
-        <div class="router-outlet">
-            <router-outlet></router-outlet>
-        </div>
-    </div>
-</div>  -->
