+++ src/app/layout/sidebar/sidebar.component.css
@@ -5,9 +5,9 @@
   transition: all 0.5s ease;
   position: fixed;
   z-index: 1;
-  top: 0;
+  top: 80px;
   width: 5rem;
-  height: 100vh;
+  height: calc(100vh - 80px); /* Ajuste para pantallas grandes */
 }
 
 .sidenav-collapsed {
@@ -85,13 +85,8 @@
   color: #000;
 }
 
-.active{
+.active {
   background-color: #fff;
-  .sidenav-link-icon,
-  .sidenav-link-text{
-    
-    color: black;
-  }
 }
 
 .logout-container {
@@ -100,7 +95,7 @@
   padding: 0.938rem 0.938rem 0 0.938rem;
   width: 100%;
   position: absolute;
-  bottom: 1.5rem; /* Ajuste para subir el botón desde la parte inferior */
+  bottom: 1.5rem;
 }
 
 .logout-button {
@@ -134,16 +129,17 @@
     width: 100%;
     height: 5rem;
     display: flex;
-    justify-content: space-around;
+    justify-content: center; /* Centra el contenido del sidebar */
     align-items: center;
     flex-direction: row;
+    padding: 0;
   }
 
   .sidenav-nav {
     position: static;
     transform: none;
     display: flex;
-    flex-direction: row;
+    justify-content: center;
     width: auto;
   }
 
@@ -153,6 +149,8 @@
 
   .sidenav-nav-link {
     height: auto;
+    display: flex;
+    align-items: center;
   }
   
   .logout-container {
@@ -162,6 +160,6 @@
   }
 
   .logo-container {
-    display: none; /* Ocultar el logo en dispositivos móviles si es necesario */
+    display: none; /* Ocultar el logo en dispositivos móviles */
   }
 }
