+++ src/app/pages/home-page/home-page.component.html
@@ -1,32 +1,37 @@
+<div class="wrapper">
+  <!-- Header -->
+  <header class="right-section">
+    <app-header></app-header>
+  </header>
 
-
-<app-header></app-header>
-<app-sidebar></app-sidebar>
-
-
-<div class="home-container">
-
-  <div *ngIf="role === 'usuario-invitado'">
-    <h2>Bienvenido a nuestra Página</h2>
-    <p>Por favor, inicia sesión para más funcionalidades.</p>
+  <!-- Sidebar -->
+  <div  [ngClass]="isSideNavCollapsed ? 'sidenav-collapsed' : ''">
+    <app-sidebar (onToggleSideNav)="onToggleSideNav($event)"></app-sidebar>
   </div>
 
-  <div *ngIf="role === 'administrador'">
-    <h2>Bienvenido, Administrador</h2>
-    <p>Tienes acceso a todas las funcionalidades.</p>
-    <app-admin-page></app-admin-page>
-  </div>
+  <!-- Contenido principal -->
+  <div [ngClass]="isSideNavCollapsed ? 'content-collapsed' : 'content-expanded'">
+    <div class="home-container">
+      <div *ngIf="role === 'usuario-invitado'">
+        <h2>Bienvenido a nuestra Página</h2>
+        <p>Por favor, inicia sesión para más funcionalidades.</p>
+      </div>
 
-  <div *ngIf="role === 'moderador'">
-    <h2>Bienvenido, Moderador</h2>
-    <p>Puedes anunciar actividades deportivas.</p>
-  </div>
+      <div *ngIf="role === 'administrador'">
+        <h2>Bienvenido, Administrador</h2>
+        <p>Tienes acceso a todas las funcionalidades.</p>
+        <app-admin-page></app-admin-page>
+      </div>
 
-  <div *ngIf="role === 'usuario-regular'">
-    <h2>Bienvenido, Usuario Regular</h2>
-    <p>Puedes agendar y marcar lugares favoritos.</p>
-  </div>
+      <div *ngIf="role === 'moderador'">
+        <h2>Bienvenido, Moderador</h2>
+        <p>Puedes anunciar actividades deportivas.</p>
+      </div>
 
+      <div *ngIf="role === 'usuario-regular'">
+        <h2>Bienvenido, Usuario Regular</h2>
+        <p>Puedes agendar y marcar lugares favoritos.</p>
+      </div>
+    </div>
+  </div>
 </div>
-
-<!-- <button (click)="logout()">Cerrar sesión</button> -->
