--- src/app/pages/settings/settings.component.html
@@ -1,1 +1,0 @@
-<p>settings works!</p>
