--- src/app/pages/settings/settings.component.ts
@@ -1,12 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'settings',
-  standalone: true,
-  imports: [],
-  templateUrl: './settings.component.html',
-  styleUrl: './settings.component.css'
-})
-export class SettingsComponent {
-
-}
