--- src/app/pages/user-manager/user-manager.component.html
@@ -1,1 +1,0 @@
-<p>user-manager works!</p>
