--- src/app/pages/user-manager/user-manager.component.ts
@@ -1,12 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'user-manager',
-  standalone: true,
-  imports: [],
-  templateUrl: './user-manager.component.html',
-  styleUrl: './user-manager.component.css'
-})
-export class UserManagerComponent {
-
-}
